Restructure file directories
