+++ thavalon-webapp/src/components/Navbar.tsx
@@ -2,7 +2,7 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import React, {MouseEvent} from 'react';
 import {NavLink} from 'react-router-dom';
 import {faBars} from '@fortawesome/free-solid-svg-icons';
-import "./Navbar.scss";
+import "../styles/Navbar.scss";
 
 interface NavbarProps {
     loggedIn: boolean
