+++ thavalon-webapp/src/styles/Modal.scss
@@ -13,6 +13,11 @@
     align-items: stretch;
     flex-wrap: wrap;
     font-size: 1.2rem;
+    width: 30rem;
+
+    @media only screen and (max-width: 600px) {
+        width: 50vw;
+    }
 }
 
 .Overlay {
